Update index.md
Added info including a link to the translation platform "Hosted Weblate" to inform that the app is translated into different languages.

This way users will know that the app is translated in other languages and may help that more users translate or improve translations of the app in their lanaguage.

Please make the link with the same blue color as when hovering over the links in the footer in the "Legal" section and ensure that the link is visible for users without hovering over the link. If necessary correct my code.